# rest/swaggerui/swagger.yaml
swagger: "2.0"
info:
  version: 0.1.0
  title: Openstack Big Data tools
  description: "Данная система предоставляет REST API для создания кластеров с заданным списком зарегестрированных сервисов в облачной среде OpenStack."
  license:
    name: License Here
  #    url:
  contact:
    email: modis@ispras.ru
host: host.here.com
basePath: /home
schemes:
  - http
  - https
...
paths:
...
  /configs/{serviceType}/versions:
    get:
      tags:
        - configs
      summary: Возвращает информацию о доступных в системе версиях указанного сервиса.
      description: Возвращает информацию о доступных в системе версиях указанного сервиса.
      operationId: configVersionsList
      parameters:
        - name: view
          description: "Полнота описания. Возможные значения: full (полное), summary (краткое). Значение по умолчанию - summary"
          type: string
          default: summary
          in: query
        - name: serviceType
          description: "Тип сервиса."
          in: path
          type: string
          required: true
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/ServiceConfigVersions'

    post:
      tags:
        - configs
      summary: Создание новой версии указанного типа сервиса.
      description: Создание новой версии указанного типа сервиса.
      operationId: configVersionCreate
      parameters:
        - name: serviceType
          description: "Тип сервиса."
          in: path
          type: string
          required: true
        - name: serviceConfigVersion
          in: body
          schema:
            $ref: '#/definitions/ServiceConfigVersion'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/ServiceConfigVersion'
...
definitions:
...
  ServiceConfigVersion:
    type: object
    example:
      {

      }
  ServiceConfigVersions:
    type: object
    example:
      {

      }